+++ Public/Set-IDIDevice.ps1
@@ -32,21 +32,6 @@ function Set-IDIDevice{
     }elseif(!$IDIDevice){
         Write-Warning "No device specified."
     }else{
-        $id = $IDIDevice.id
-        $Notesonly = $($IDIDevice[0] | Convertto-Json) | Convertfrom-Json
-        $RefProperties = (Get-noneIDIReference).PSObject.Properties.Name
-        foreach($Property in $RefProperties){
-            $Notesonly[0].PSObject.Properties.Remove("$Property")
-        } 
-        Write-Verbose "Update Device $id"
-        Write-Verbose "  Properties: $($Notesonly.PSObject.Properties.Name)"
-
-        #Update the notes of the device
-        $Note_json = $Notesonly | Convertto-Json
-        $Json = @{ "notes" = "$Note_json" } 
-        $uri = "https://graph.microsoft.com/beta/deviceManagement/managedDevices('$id')"
-
-        $MSGraphRequest =  Invoke-MSGraphRequest -Url $uri -HttpMethod PATCH -Content $Json
-        Write-Verbose "$MSGraphRequest"
+        Set-IDIDeviceNotes -IDIDevice $IDIDevice
     }
 }