+++ ReleaseNotes.md
@@ -1,4 +1,17 @@
 # Release notes for IntuneDeviceInventory module
+## 0.1.2
+Added functions:
+- Get-IDIDeviceNotes
+- Invoke-IDIDeviceBitLockerRotation
+- Invoke-IDIDeviceDefenderScan
+- Invoke-IDIDeviceDefenderSignatures
+- Invoke-IDIDeviceRestart
+- Invoke-IDIDeviceSync (changed from Sync-IDIDevice)
+- Set-IDIDeviceNotes
+- Start-IDI
 
-## 0.1
+## 0.1.1
+Fixes and performance optimisation
+
+## 0.1.0
 - Initial release, se README.md for documentation.